refactor(src): remove unused services, contributor proposal and vesting section
