--- src/components/UserPage/VestingContractsSection/index.ts
@@ -1,1 +1,0 @@
-export { default as VestingContractsSection } from './VestingContractsSection';
