+++ src/components/UserPage/index.ts
@@ -1,2 +1,1 @@
 export * from './styled';
-export * from './VestingContractsSection';
