--- src/components/UserVestingInfoModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './UserVestingInfoModal';
