+++ src/pages/SubmitProposal.tsx
@@ -3,7 +3,6 @@ import { observer } from 'mobx-react';
 import { Switch, Route, useRouteMatch } from 'react-router-dom';
 
 import NewProposalPage from './ProposalSubmission/Custom';
-import { ContributorProposalPage } from './ProposalSubmission/ContributorProposal';
 
 export const SubmitProposalPage = observer(() => {
   let match = useRouteMatch();
@@ -12,9 +11,6 @@ export const SubmitProposalPage = observer(() => {
     <Switch>
       <Route path={`${match.path}/custom`}>
         <NewProposalPage />{' '}
-      </Route>
-      <Route path={`${match.path}/contributor`}>
-        <ContributorProposalPage />{' '}
       </Route>
     </Switch>
   );
